+++ api/v2/v2.go
@@ -2,8 +2,9 @@ package v2
 
 import (
 	"context"
-	gw "github.com/MinterTeam/minter-go-node/api/v2/api_pb"
 	"github.com/MinterTeam/minter-go-node/api/v2/service"
+	gw "github.com/MinterTeam/node-grpc-gateway/api_pb"
+	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/tmc/grpc-websocket-proxy/wsproxy"
 	"golang.org/x/sync/errgroup"
@@ -18,8 +19,13 @@ func Run(srv *service.Service, addrGRPC, addrApi string) error {
 		return err
 	}
 
-	grpcServer := grpc.NewServer()
+	grpcServer := grpc.NewServer(
+		grpc.StreamInterceptor(grpc_prometheus.StreamServerInterceptor),
+		grpc.UnaryInterceptor(grpc_prometheus.UnaryServerInterceptor),
+	)
 	gw.RegisterApiServiceServer(grpcServer, srv)
+	grpc_prometheus.Register(grpcServer)
+
 	var group errgroup.Group
 
 	group.Go(func() error {
@@ -30,7 +36,10 @@ func Run(srv *service.Service, addrGRPC, addrApi string) error {
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
 	mux := runtime.NewServeMux(runtime.WithMarshalerOption(runtime.MIMEWildcard, &runtime.JSONPb{OrigName: true, EmitDefaults: true}))
-	opts := []grpc.DialOption{grpc.WithInsecure(), grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(50000000))}
+	opts := []grpc.DialOption{
+		grpc.WithInsecure(),
+		grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(50000000)),
+	}
 
 	group.Go(func() error {
 		return gw.RegisterApiServiceHandlerFromEndpoint(ctx, mux, addrGRPC, opts)