+++ cmd/minter/main.go
@@ -1,12 +1,22 @@
 package main
 
 import (
+	"context"
 	"github.com/MinterTeam/minter-go-node/cmd/minter/cmd"
 	"github.com/MinterTeam/minter-go-node/cmd/utils"
+	"github.com/tendermint/tendermint/libs/log"
+	tmos "github.com/tendermint/tendermint/libs/os"
+	"os"
 )
 
 func main() {
 	rootCmd := cmd.RootCmd
+
+	ctx, cancel := context.WithCancel(context.Background())
+
+	tmos.TrapSignal(log.NewTMLogger(os.Stdout).With("module", "consensus"), func() {
+		cancel()
+	})
 
 	rootCmd.AddCommand(
 		cmd.RunNode,
@@ -15,14 +25,20 @@ func main() {
 		cmd.ManagerCommand,
 		cmd.ManagerConsole,
 		cmd.VerifyGenesis,
-		cmd.Version)
+		cmd.Version,
+		cmd.ExportCommand,
+	)
 
 	rootCmd.PersistentFlags().StringVar(&utils.MinterHome, "home-dir", "", "base dir (default is $HOME/.minter)")
 	rootCmd.PersistentFlags().StringVar(&utils.MinterConfig, "config", "", "path to config (default is $(home-dir)/config/config.toml)")
 	rootCmd.PersistentFlags().Bool("pprof", false, "enable pprof")
 	rootCmd.PersistentFlags().String("pprof-addr", "0.0.0.0:6060", "pprof listen addr")
 
-	if err := rootCmd.Execute(); err != nil {
+	cmd.ExportCommand.Flags().Uint64("height", 0, "export height")
+	cmd.ExportCommand.Flags().String("chain_id", "", "export chain id")
+	cmd.ExportCommand.Flags().Duration("genesis_time", 0, "export height")
+
+	if err := rootCmd.ExecuteContext(ctx); err != nil {
 		panic(err)
 	}
 }